tensor_move 整数溢出 coalesce_sparse 存在除零错误风险整改



# message auto-generated for no-merge-commit merge:
!278 merge modify_math into master

tensor_move 整数溢出 coalesce_sparse 存在除零错误风险整改

Created-by: zhuzemao
Commit-by: zhuzemao
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-math!278

+++ conversion/tensor_move/op_host/op_api/tensor_move.cpp
@@ -76,7 +76,7 @@ const aclTensor* TensorMoveAiCpu(const aclTensor* x, const aclTensor* y, aclOpEx
 
 bool IsCopyNpuToNpu(const aclTensor* x)
 {
-    auto dataSize = x->GetViewShape().GetShapeSize() * op::TypeSize(x->GetDataType());
+    auto dataSize = static_cast<uint64_t>(x->GetViewShape().GetShapeSize()) * op::TypeSize(x->GetDataType());
     auto socVersion = op::GetCurrentPlatformInfo().GetSocVersion();
     if (socVersion == op::SocVersion::ASCEND910B || socVersion == op::SocVersion::ASCEND910_93) {
         return static_cast<int64_t>(dataSize) <= DATA_LIMIT_910B;
